--- libraries/Scheduling/src/Assimalign.Extensions.Scheduling.Crontab/ExceptionProvider.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Assimalign.Extensions.Scheduling;
-
-
-public delegate Exception ExceptionProvider();
