--- libraries/Scheduling/src/Assimalign.Extensions.Scheduling.Crontab/Internal/StringSeparatorStock.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Assimalign.Extensions.Scheduling;
-
-internal static class StringSeparatorStock
-{
-    public static readonly char[] Space = { ' ' };
-    public static readonly char[] Comma = { ',' };
-}
